added big decimal valueOf

+++ src/main/java/com/example/swe_finanzmanager/backend/persistence/XMLReader.java
@@ -118,7 +118,7 @@ public class XMLReader {
 				(Date) Date.valueOf(element.getElementsByTagName("datum").item(0).getTextContent()),
 				(Nutzer) sp.getObject(ClassType.NUTZER,
 						element.getElementsByTagName("ersteller").item(0).getTextContent()),
-				(Konto) new Konto(0, null, path, path, 0, path),//////Kontos sind noch nicht geladen-> wird später eingefüllt
+				(Konto) new Konto(BigDecimal.valueOf(0), null, path, path, 0, path),//////Kontos sind noch nicht geladen-> wird später eingefüllt
 				(String) element.getElementsByTagName("beschreibung").item(0).getTextContent(),
 				(String) element.getElementsByTagName("titel").item(0).getTextContent());
